+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### Changed
 - Update pluginIcon.svg
 
+### Fixed
+- Subclasses resolving
+
 ## [0.1.0]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)