+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # BDLH Changelog
 
 ## [Unreleased]
+
+## [0.2.0]
 ### Added
 - Navigation for class signature
 
@@ -17,4 +19,3 @@
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
 - Partial highlighting and navigation for method and field signature
-